Move everything from cpp files to header files in namespace bamit and update cmake.
