+++ src/CMakeLists.txt
@@ -1,10 +1,9 @@
 cmake_minimum_required (VERSION 3.8)
 
 # An object library (without main) to be used in multiple targets.
-add_library ("${PROJECT_NAME}_lib" STATIC IntervalNode.cpp
-                                          Record.cpp)
-target_link_libraries ("${PROJECT_NAME}_lib" PUBLIC seqan3::seqan3)
-target_include_directories ("${PROJECT_NAME}_lib" PUBLIC ../include)
+add_library ("${PROJECT_NAME}_lib" INTERFACE)
+target_link_libraries ("${PROJECT_NAME}_lib" INTERFACE seqan3::seqan3)
+target_include_directories ("${PROJECT_NAME}_lib" INTERFACE ../include)
 
 add_executable ("${PROJECT_NAME}" main.cpp)
 target_link_libraries ("${PROJECT_NAME}" "${PROJECT_NAME}_lib")
